+++ src/components/request/DateRangeInput.tsx
@@ -1,27 +1,20 @@
 import React, { useState } from "react";
 import { DateRange } from "react-date-range-ts";
 import { ko } from "date-fns/locale";
+import { useRecoilState } from "recoil";
+import { requestDateRangeAtom } from "../../state/atoms";
 import "react-date-range-ts/dist/styles.css";
 import "./styles/request-calendar-custom.css";
-
 import { StyledDateRangeInput, StyledCalendarContainer } from "./styles/dateRangeInputStyle";
 import { convertDateToString, convertDateWithDots, convertStringToDate } from "../../shared/utils/dateHandlers";
 
-type DateRangeValues = {
-  startAt: string;
-  endAt: string;
-}
+const DateRangeInput = () => {
+  const [dateRange, setDateRange] = useRecoilState(requestDateRangeAtom);
 
-type InputProps = {
-  value: DateRangeValues;
-  setValue: React.Dispatch<React.SetStateAction<DateRangeValues>>;
-};
-
-const DateRangeInput = ({ value, setValue }: InputProps) => {
   const [isCalendarOpen, setCalendarOpen] = useState(false);
   const [selectedRange, setSelectedRange] = useState({
-    startDate: value.startAt ? convertStringToDate(value.startAt) : new Date(),
-    endDate: value.endAt ? convertStringToDate(value.endAt) : new Date(),
+    startDate: dateRange.startAt ? convertStringToDate(dateRange.startAt) : new Date(),
+    endDate: dateRange.endAt ? convertStringToDate(dateRange.endAt) : new Date(),
     key: "selection"
   });
 
@@ -30,7 +23,7 @@ const DateRangeInput = ({ value, setValue }: InputProps) => {
   };
 
   const handleClickSubmit = () => {
-    setValue({
+    setDateRange({
       startAt: convertDateToString(selectedRange.startDate),
       endAt: convertDateToString(selectedRange.endDate)
     });
@@ -43,7 +36,7 @@ const DateRangeInput = ({ value, setValue }: InputProps) => {
         <span>이벤트 기간</span>
         <button type="button" className="calendarOpenInput" onClick={() => setCalendarOpen(!isCalendarOpen)} >
           <input disabled id="dateRange" type="text" placeholder="날짜 선택하기"
-                 value={(value.startAt && value.endAt) ? `${convertDateWithDots(value.startAt)} - ${convertDateWithDots(value.endAt)}` : ""} />
+                 value={(dateRange.startAt && dateRange.endAt) ? `${convertDateWithDots(dateRange.startAt)} - ${convertDateWithDots(dateRange.endAt)}` : ""} />
           <i id="calendar"/>
         </button>
       </div>
